-) Added new mc15 interface for IsolationSelectionTool in:
      Root/ElectronSelector.cxx

and saved a decoration for electron that pass isolation.

-) Added new class:

      xAODAnaHelpers/ParticlePIDManager.h

to manage different WPs for Particle Identification. It is used as a member of the *Selector modules, and it  will configure only tools with WP equal or tighter than the one selected,
to allow selection on the required WP, and decoration for the tighter WPs. For now, done only for electrons (LH based ID).

-) The isolation and PID decorations are now added as branches of the output tree, configurable via switches defined in HelperClasses.

-) Cleaned up the electron*.config files to make them more readable.