Edits for DataScouting compatibility (#1172)
* allow multiple trigger jet collections in TreeAlgo

* add multiple trigJet_detailStr to TreeAlgo as for reco

* add config option to sort L1 RoIs by et8x8

* fix compile warning

change int to unsigned int

* add config option to sort L1 RoIs by et8x8

* allEventsCBK protection
The stream is sometimes unknown in DataScouting AODs, meaning that allEventsCBK is not set. This change adds protection against this, setting the relevant values to 0 in such a situation along with a warning.

* Option to ignore lack of Primary Vertices
DataScouting does not have tracks or vertices, and so PV checks were causing crashes.
Added an option to skip PV checks, and then downgrade a warning to debug in HelperFunctions::getPrimaryVertexLocation to avoid spam.
There are perhaps better ways to deal with the latter issue that will retain the warning for most users.

* Add new jet cleaning infoSwitch
DataScouting jets do not have the summary cleaning variables - eg clean_passLooseBad
Added a new infoSwitch option, "cleanNoSumm", that doesn't request these variables and hence avoids
errors being printed for every jet.

* allEventsCBK protection
The stream is sometimes unknown in DataScouting AODs, meaning that allEventsCBK is not set. This change adds protection against this, setting the relevant values to 0 in such a situation along with a warning.

* Option to ignore lack of Primary Vertices
DataScouting does not have tracks or vertices, and so PV checks were causing crashes.
Added an option to skip PV checks, and then downgrade a warning to debug in HelperFunctions::getPrimaryVertexLocation to avoid spam.
There are perhaps better ways to deal with the latter issue that will retain the warning for most users.

* Add new jet cleaning infoSwitch
DataScouting jets do not have the summary cleaning variables - eg clean_passLooseBad
Added a new infoSwitch option, "cleanNoSumm", that doesn't request these variables and hence avoids
errors being printed for every jet.