Update Root/BJetEfficiencyCorrector.cxx: add/remove new/old taggers
Taggers have been updated:
https://gitlab.cern.ch/atlas/athena/commit/e36de3cc8d4b65e065e3f09c3e45cec64bd64442
and use of deprecated ones now result in `StatusCode::FAILURE`.

This commit removes from the list of "known" xAH taggers and adds the
new ones. N.B. as of the date of this commit, weights are not
available in old(er) derivations for the new taggers!