BasicEventSelection::finalize(): Remove TDT from ToolStore (#1295)
Simply calling finalize of the TDT does some nasty things under the hood in the TDT itself. So say if you have two algs (say, `xAH` and `SUSYTools` -- don't hate. Let's be friends.) grabbing the TDT from the tool store, if they both try to finalize, it barfs a bit. The proposed change ensures that after calling `finalize()` on the TDT, it's removed from the `ToolStore` so that nobody else can try to call `finalize()` again.