Grabbing TEvent and TStore in BES::initialize() (#1314)
Because of what looks like a change in execution order in 21.2.61, the TEvent and TStore being grabbed in the `fileExecute()` doesn't make it available for the `initialize()`. So proposing that these are also grabbed in the `initialize()` to protect against this. Maybe a bit redundant, but would make this less sensitive to function call order. Can we do something like this?

Thanks!