Muon container primary vertex check (#1401)
* Truth particle tree name updated

The default name of the output truth particle tree was "xAH_truth" in AddTruthUser() and AddTruthUser() but "truth_particle" in FillTruth(). This caused a crash when running with default output truth tree names. AddTruthUser() default was updated to "xAH_truth" to prevent crash.

* Muon container primary vertex check

In MuonContainer.cxx, the primary vertex is accessed without checking if it is a null pointer. This caused infrequent but fatal failures when filling muon trackparams. A simple check is added so if the primary vertex is null, m_trkz0 is filled with -999.